requirements.txt
aiogram>=3.0.0,<4.0.0
openai>=1.0.0,<2.0.0
supabase>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
apscheduler>=3.10.0,<4.0.0
